Make ExecutionContext.Clone() copy the host.
Without this, writing to the host would fail with a NullReference. But is this really the right approach? I don't know enough about how ExecutionContexts should work.

Also, relying on Clone() to get updated with every new field is error-prone. I hope the future includes a better approach that is less error-prone.