REFACTORING: ConsoleView shouldn't automatically show a prompt
Before, ConsoleView *always* showed the default prompt ("> ") in its
ctor. That meant you couldn't override the prompt the first time around.

After the first time, it's up to the client of ConsoleView to decide
when to show the prompt. So there was an assymetry here.

I changed ConsoleView to not show the prompt the first time around. The
client now has sole responsibility for deciding when to show the prompt.