Expose token patterns as `RegexBasedTerminal` instead of a `string`
This is basically a refactoring to move away from the primitive type,
but it also has the important effect of making the `Terminal`s in to
singletons, so they can be reused and compared as needed, and so we
have a good name to pass in to the `RegexBasedTerminal`.

`RegexBasedTerminal` takes both a rule name and a pattern in its ctor.
The rule name is a duplicate of the field name, but DRY. To make the
reflection work out, I store the pattern in one field, and then
construct the object at runtime.

The current code only works if all the terminals are
`RegexBasedTerminal`s. If we decide we want to use another type of
terminal, we'll need to deal with that.