Use Unicode escape codes, not actual characters, in the regexes.
This makes reading the resulting regex a little easier. For example, a
newline won't look like a blank line, but like "\u000A\u000D".

Since escaping strings is a little tough to read, I made any string that
uses a backslash in to a verbatim string literal, for less escaping.