IRONY: When creating unnamed Terminals, mark them as transient.
When a language grammar has an optional element:

    FOO ::= BAR BAZ_opt

a common idom in Irony is to write:

    Foo.Rule = Bar + (Baz | Empty)

BNF is strict that you can have an or-ed list of ands, but not an and-ed
list of ors. So Irony rewrites the rule like this:

    Unnamed0.Rule = Baz | Empty
    Foo.Rule = Bar + Unnamed0

that is, Irony invents a new rule, just for this purpose. That's a problem
if you want to create an AST, as you have no way to tell Irony what AST
node to create for this invented rule.

With my change, these invented rules are marked "transient", which means
they won't appear in the final parse tree, and won't affect the AST.