Name regex groups in Terminals
this makes it easy to extract a specific element from the matched text.

For example, a verbatim string literal can be delimited with 4 different
types of quotes, and the rest of Pash shouldn't have to worry about
figuring that out. Now you can just do a regex match & pull out groups
by name.

This should also be useful when making sense of more complex tokens,
such as expandable strings.