Shift/Reduce errors, and more.
I examined the parser under Irony.GrammarExplorer for the first time (as advised by the creator of Irony). This found many Shift/Reduce conflicts, and a few Reduce/Reduce conflicts.

Most of these conflicts are ignorable, as they only relate to the newlines in the grammar. For example:

    Foo.Rule = X + (newlines | Empty) + (Bar | Empty) + (newlines | Empty) + Y;

When Bar is not present, a newline could match either instance of `(newlines | Empty)`. Newlines are ignored in the semantics, so this isn't important, but it is still an error in the grammar.

However, a few of the conflicts were due to ligitimate bugs in the grammar. I have fixed or disabled them. At a later point, we should make an effort to address all remaining shift/reduce and reduce/reduce errors as reported by Irony, and implement an automated test to keep the nuimber at 0.

Also:

- Terminals are no longer statics, so no more static ctor. The regex patterns are still `const strings`, but the `Terminal` objects are now instance fields. An upside is that referencing terminals and nonterminals now uses the same syntax. A downside is that writing unit tests for the terminals is now impossible, without changing accessibility to access the terminal fields from the tests. That's OK with me, because the higher level AST tests give good coverage that's easy to write.

- Built out Ast for a few more items, and normalized some of the Ast code. It will be fun to flush out building the rest of the Ast

List of squashed commits:

	230cbb6 comment linking disabled attributes to Issue:11
	95fbac7 Merge branch 'master' into work
	7bcf3c1 shift/reduce conflict in array_literal_expressions
	9310d0c Fix reduce/reduce error around redirection.
	c3d909e Remove shift/reduce error around attribute_list
	b8abc48 Disable command_module - see https://github.com/JayBazuzi/Pash2/issues/8
	366d06f Issue:9 - Disable grammar that require more delecate handling of whitespace.
	f71fb37 Disable some rules that require more delecate handling of whitespace.
	d3d88b4 Remove bad comment
	4374fda Disable line continuation with `\r\n
	31fd82d Fix Shift/Reduce conflict in trap_statement.
	073e81b Fix Shift/Reduce conflict in param_block.
	8a51588 Remove semantically optional `else`
	375b435 Use MakePlusRule for pipeline tails.
	ffff398 make terminals non-static
	5e89502 Normalize ast builders:
	127f536 Rename ParseTreeNode parameters to parseTreeNode; move VerifyTerm to the right place.
	d21d555 Expand BuildStatementAst
	0bc6c42 Build AST for member access (`[System.Int32]::MaxValue`)