Make root of parse tree always `script_block`
The language spec has 4 possible parse tree roots:

- script_file
- module_file
- interactive_input
- data_file

The first 3 are always `script_block`. I've only been using
`interactive_input` thus far, even if the input is not interactive. The
others violate YAGNI and the code is a little confusing. For now, I'm
going directly to `script_block` as the root.