Eliminate Pash.Configuration.
As far as I can tell, the purpose of this project was to make it easy
to declare the default config of a host - prompt, aliases, etc., using XML
in App.config.

I think a better way to do this is with a .ps1 script, rather than in XML.
Most of the content was moved by previous commits. The remainders were:

- List of snapin providers to load. This is now hard-coded in
`CommandManager.cs`, which is bad (although not the worst problem at the
moment.)

- The $false and $NestedPromptLevel variables. But these didn't actually
work, anyway. All in good time.

One positive consequence of this change is that the build is now
warning-free.