IRONY: Rename files for case sensitivity
These files were named `RegEx...` on disk, but `Regex...` in the project
file. That's fine on Windows, but we also want to be able to build under
MonoDevelop on Linux. On Linux, file names are case sensitive, so the
build failed.

Instead of renaming the file, we could have modified the project file to
use `RegEx...`, but this way matches the name of the class, and the way
`Regex` is spelled in the .NET Base Class Libraries.

This fix merged from https://github.com/b333z/Pash2/commit/ea10fa5. Thanks!