Don't use wildcards in project files.
I like wildcards in project files a lot. Without wildcards, it's easy to
create a file on disk but forget to add it to the project, and end up
with silent failures.

Visual Studio's support for wildcards is not perfect, but I was prepared
to tolerate it.

However, MonoDevelop's support for wildcards is pretty broken, and I want
people to be able to work on Linux.

So I'm going back to no wildcards. I did this by going to each project in
VS's Solution Explorer, removing a single file from the project, then
adding it back in.

This undoes commit ee6fcbbb63fb1d2dff1c9371ce499a3629ded8fe.