Mark Solution as VS2010.
VS2012 is the first version of Visual Studio that is compatible with
projects and solutions created in earlier versions. It is also supposed to
save those projects in the older, compatible version. That quietly broke
with commit 3453e3226cecb2a50e95b510ec0e70ea0c692696.

It's particularly important because MonoDevelop is not compatible with VS2012
projects/solutions yet.