REFACTOR: Use reflection to find `AstVisitor` dispatch method.
While doing this work, I realized there are `AstVisitor` methods that take
both a specific and general type. For example, a
`StringConstantExpressionAst` can be sent to
`VisitStringConstantExpression` or `VisitConstantExpression` (or both!).

I don't know what the official PowerShell behavior is, but for now we
don't rely on it, so I'm ignoring it.