Use `Explicit` instead of `Ignore` for failing tests.
Some unit tests are intended to decribe what *should* happen, but they
fail because of a bug or missing functionality. To help track that goal
I kept the test around but marked it as `Ignore`, so NUnit wouldn't run
it.

As I make progress on new functionality, I want to check which
previously-failing tests are now passing. By marking a test as `Explicit`,
I can easily execute the test in the NUnit GUI to see if it now passes.

I also found that a couple of tests now pass, so I enabled them.