REFACTOR: Use LINQ
Thanks to LINQ, we no longer need to iterate through collections to
convert them to `Array`

Note that returning `Array` from this method is a little weird - usually
you'd return `object[]` or something. There's only one caller, and it
doesn't use the return value, so this code doesn't matter (yet).