Fix filename starting with a space.
If you use `git` to clone this repo, things work OK. But if you
download a zip of the source and use Windows Explorer to unzip it, the
leading space is stripped out. Since the project file expects the space,
we get a build error.

The leading space in this filename is from the original code. See http://pash.svn.sourceforge.net/viewvc/pash/trunk/Pash/Pash.System.Management/Automation/%20ValidateEnumeratedArgumentsAttribute.cs?view=log

Interestingly, Visual Studio's Solution Explorer was unable to handle a
rename to remove a leading space. I renamed elsewhere, then removed the
file from the project and added it back in.

Thanks to .B ekiM for reporting this issue.
Windows