Remove never-run code in WriteHost.
This code triggers a compile warning (which is treated as an error) when using MonoDevelop.

The condition is bogus - an object can never satisfy `is Enumerable` because `Enumerable` is a `static class`. So it will always be false. I probably intended to write `is IEnumerable` here, but that causes tests to fail. So, yanking this broken code.

I am also opening Issue #20 to make sure we come back and think about this carefully at some point.