Fix warnings as errors in MonoDevelop
It seems MonoDevelop is unable to properly process
Condition attributes in project files.  As a result the
Conditional TreatWarningAsErrors added in SHA: e06750c cause
MonoDevelop to always set TreatWarningAsErrors to true.

This commit switches the order of the lines as MonoDevelop seems
to process the first instance and ignore the second, changing the
behaviour back to TreatWarningsAsErrors=false