Verify that Mono isn't busted.
There was a bug in Mono's method overload resolution that I found last
August. The Mono team fixed it promptly, but if you don't have the fix on
your machine, it will cause the Irony parser to behave badly. This is the
underlying cause of Issue #24.

The fixed Mono isn't in a patched Ubuntu 12.10, which is used by
Travis-CI.

Here I create a unit test for this Mono bug, and modified .travis.yml to run
that test before doing anything else.