REFACTOR: Avoid using `ToString()` as part of an class's contract.
`ToString()` is great for diagnostics but it's a bit dodgy for other
uses. For example, it can't be localized, so it shouldn't normally be
showing in the UI.

One that I would have liked to fix, but did not, is in
`GetMemberCommand.ProcessRecord()`:

    members.Add(new MemberDefinition(fullName, info.Name, info.MemberType, info.ToString()));

It was beyond my understanding to pick a good fix, so I let it go for now.