`New-Object` on something not in `mscorlib`.
`Type.GetType()` looks in the current assembly and in `mscorlib`. If you
want something different, you have to use `Assembly.Load().GetType()`.

In this checkin I also create a Spikes project. It's full of NUnit tests,
but it's not about learning. It's a place to learn by experimentation. The
experiments can be saved, as an easy way of recording what we learned.