Replace `| Empty` with `_opt` rules.
If you use write:

    Foo.Rule = Bar + (Baz | Empty);

Then Irony will create an "unnamed" `NonTerminal` for the `Baz | Empty`
part. Seeing those in GrammarExplorer made it hard to figure out what I
was looking at.

I replaced each `Baz | Empty` with `Baz_opt`, which is defined as:

    Baz_opt = Baz | Empty;

so the semantics are the same, but now the elements are named.

Also, marked them as `transient` so they don't appear in the final parse
tree. It might be better not to do this, so the AstBuilder can more easily
deal with it, but that would require updating the AstBuilder before I
could commit these changes.