Add a bunch of parser tests.
The point here is to establish a baseline for the parser, before changing the grammar. Some tests failed, so they're marked "Explicit". Over time, these should get fixed.

To create the tests, I read through the language spec looking at the "Examples" subsections, and turned them each in to a test case.

I got tired after 7.1.6, so that's where I stopped for now.

Also includes a handful of hand-written tests.