Add a minimal copyright header to all source files.
While we own copyright to our work by default, it's easier to enforce
copyright if we explicitly state it in each source file. That would mean
each file gets a statement of copyright for each contributor with the
year(s) of the contributions listed explicitly. That's very tedious, and
therefore error-prone.

License must be explicitly granted, and the more explicit you are, the
easier it is to enforce. The belt-and-suspenders approach is to put a
LICENSE file in the root directory *and* a comment header at the top of
every source file (with the full license in the case of MIT/BSD, and an
few short paragraphs in the case of GPL). Personally, I find it very
annoying when I open a source file and the first screenful is boilerplate
legalese that I want to ignore. And with well-factored code (small
classes), the legalese can be bigger than the code - a sign that
something is very wrong.

For precedent, I looked at several other open source projects. Most
didn't bother with copyright/license headers of any kind. Igor didn't
bother with comment headers with his original release of Pash, either.

I'm not worried about people "stealing" Pash code, since we're not really
innovating here.

I decided to turn the dial ever so slightly towards more legalese, by
adding a single line indicating copyright and license in every source
file. I hope this is a good balance between legal protection and
developer comfort.

TODO: Automate verifying that this is in place, new files get added.