Remove grammar from comments in `AstBuilder.cs`.
The comments represent the grammar as described in Microsoft's PowerShell
Language Specification document, but the implemented grammar diverges from
that document in several ways. Keeping the comments in `AstBuilder.cs` in
sync with the actual implemented grammar in `PowerShellGrammar.cs`
voilates DRY, and was getting out of sync. Removing it.