REFACTOR: Rename many "Unnamed" parse tree nodes.
In Irony, if you write:

     A.Rule = B + (C | D)

it will automatically generate a term with a name like `Unnamed9` for
the `C | D` part. These are opaque when analyzing the grammar in
Irony.GrammarExplorer. To make that task easier, I am introducing some
named terms to replace them. There are now only 3 unnamed terms.