Sync with Launchpad
Igor made the original release of Pash in spring of 2008. That
summer, Jonathan Ben-Joseph did some work on Pash as part of the Mono
Summer of Code. This merges in his changes.

Jonathan's release of his code on Launchpad indicated an MIT License. I
have email from Jonathan on 31 December 2012 agreeing to release that code
under BSD+GPL as well, so it aligns with Igor's original license.

@alesliehughes did the work of merging this code.