Fix parsing of `if ($true) { }`
The whitespace between the braces was confusing the parser. I can't say I
understand why.

Added a test case for a conditional with a more complex condition, I've
seen this cause bad parsing results - there are so many different parts
of the grammar at play here, and ambiguity is a possible problem.