Parsing of `-ne` expressions.
Oddly, this worked:

    1 -eq 2

but this failed to parse:

    1 -ne 2

Why?

When Irony scans for tokens, it uses roughly the following algorithm:

1. Based on the current state of the parser, which non-terminals are
allowed next?

2. Try to match each terminal to the current input.

3. If 2 terminals have the same priority, prefer one that matches
more of the input (be greedy)

4. Return the highest-priority terminal that matches

There's a problem here: if 2 matching terminals have the same priority
and match the same length of input, the match is non-deterministic.
This indicates a problem in the grammar definition, but it's one that's
easy to miss.

I discovered this problem while trying to test a comparison with `-ne`.
The parser sees it as either a command parameter or a comparison
operator, and was chosing command parameter. (When it saw `-eq`, it
picked comparison operator, again non-deterministically.)

The fix is to raise the priority of comparison operator terminals.

I added an exception in Irony for these ambiguities, to make it obvious
when it happens. This caused some tests to fail, because of an
ambiguity between generic and array type expressions. I temporarily
disabled those parts of the grammar.