Continue parsing past whitespace within a statement.
PowerShell allows whitespace within a statement:

    if ($true)     # continues parsing this statement on the next line
    {
    }

although this doesn't work if the grammar doesn't require it:

    echo ($true)    # end of statement
    {
    }

The PowerShell language specification describes this by inserting
`new_line_opt` in the grammar, wherever a newline should not be
recognized as a statement terminator, and also putting
`new_line_character` in the `statement_terminator` rule. See
commit 241c98cd77daf918cc76555dc0324817dd0514d0 for what that looks
like.

The alternative, suggested by [the maintainer of Irony] (https://www.codeplex.com/site/users/view/rivantsov),
which I have implemented here, is to process those newlines at runtime.
`new_line` is still a `statement_terminator`. Whenever the scanner
sees a `new_line`, we check whether the grammar expects as
`statement_terminator` at that location. If it does not, we replace
the `new_line` token with a non-grammar `skipped_new_line` token.

Take a peek in `config.ps`, where the prompt is now declared across
multiple lines.