Issue #12: Fix line continuation on \r\n
Previously I had overloaded `SkipWhitespace` to implement line
continuation, since the PowerShell Language Specification calls it a
kind of whitespace. However, there wasn't a way to preview `\r\n` there -
you can only preview one character at a time. I wrote an extension to
Irony to allow 2-character preview, but it hasn't been incorporated in to
the stock Irony codebase.

It's undesriable to depend on a local extension to Irony, because then the
stock Irony Grammar Explorer won't work. Plus, it makes keeping in sync
with new versions of Irony that much harder.

Irony includes a `LineContinuationTerminal` that does most of the work for
you, it only needed a few parameters set to make it do the
right thing for the PowerShell language.

I also modified the definition of `escaped_character_pattern` because it
was eating the backtick as part of a `generic_token`, instead of leaving
it for line continuation - bad!