Array indexing.
This works:

    PASH > (10..12)[1]
    11

Making this work gets us 1 step closer on Issue #11.

There is a growing trend in the code of unwrapping `PSObject`s, and that must be corrected at some point.

There is also some growing duplication around conversions, which needs to be DRYed out and brought in alignment with the spec.