Work around Mono GetDrives() bug
The results of System.IO.DriveInfo.GetDrives() is incorrect on my
system, returning a single entry with a blank name.

This commit works around the issue by catching this condition when
running on *nix platforms, and returning a single drive called root.

*nix like operating systems are less reliant on the concept of a drive
anyway, so this allows Pash to continue operation and still represent
the cwd.