Parameters with no set specified should belong to all sets
If a CmdLet has parameter sets, and a parameter has
no parameter set specified, then that parameter should
belong to all parameter sets.

This was stopping parameters from being passed correctly to a CmdLet
when it has multiple parameter sets, for parameters that had no set
specified.

More work needs to be done to the parameter logic but should
improve it slightly.

As a result the following should now work:

	Get-ChildItem -Recurse