Implement `return` for scriptblocks. Now this works:
    PASH> & { return ; 1 } ; 2
    2

I used an exception to abort the current script block. I'd rather avoid
using a exceptions in normal program flow, but I didn't see a
straightforward way to handle the way `return` can pop so much execution
state.

One downside is that the AST Visitor dispatcher, which uses reflection to
invoke the handler, unwraps and rethrows the exception. Ugly.