PathShouldEqual for other drives than C:
Previously running the tests on Windows from another drive than C: (e.g.
D:) would produce 8 test failures. Since they do work when the current
folder is on the C: drive this can be a bit vexing to debug and figure
out. Tests in general shouldn't depend on external factors like the
current folder to fail or succeed, even if they are currently a hack.