Fixed a missing `$` typo in a test case.
This was taken verbatim from the specification but funnily enough it
doesn't currently fail because instance property access isn't
implemented (e.g. `'xxx'.Length` works). It fails because `$a` isn't an
array, but a collection and has no `Length` member (but `Count` – arrays
in PowerShell always have both `Count` and `Length`, though).