Cleanup. Moved Ctrl+C handling to UI.
All "Console" related stuff is now in the LocalHostUserInterface class, where it belongs.
The rest of the application should be independent from the Console class now, s.t. the
user interface is completeley exchangable.
As Ctrl+C handling should be possible and directly handled by the user interface, some
changes in other classes have been made in order to access the currently running pipeline
from the user interface an stop it.
This behavior is better than throwing an exception in "ReadLine", because each call to ReadLine
would need to implement the "stop pipeline" behavior on its own.

Last but not least, code was removed that was duplicate/not in use/in comments and unnecessary.