VisitPipelineExpression uses Pipeline. Fixes.
The lifecycle of commands in a pipeline was changed, otherwise it didn't make sense.
This means, a command has to run "EndProcessing" before the next command processes records.
Inserted comments on how the pieline is supposed to work: ProcessRecord() as soon as the previous command
writes to the pipeline.
Also, the ProcessRecord phase is only entered if there are input objects or the command is first in the pipeline.
This is how PS does it.

Modified tests to test and work properly. Also added tests for the desired pipeline behavior which isn't implemented, yet.
So these tests are currently ignored.

Fixed the possibility to clear the $error variable.
Some further fixes.
All tests work now und Unix.