Behavior adjustments for reference tests
If a parameter set of a cmdlet is invalid, an exception
will only be thrown if it's intended to be used, not when it's loaded.
Otherwise Pash won't work correctly just because of one malformed
cmdlet.

Also, the Pipeline itself now stops working when there is a terminating error,
e.g. if a cmdlet uses ThrowTerminatingError or there is an parse error.
Catching and printing a broken pipeline is not part of the framework, but
just of the console. That's why this behavior is now in FullHost.

The ReferenceHost is now a separate class and should be used by both the
PS and Pash project.

To use the ReferenceTests with Pash, more projects/assemblies had to be added
to the project. This is an issue that should be resolved.