Better test, parameter lookup fixes, ErrorRecords
One of the reference tests actually tests if a parameter set is created
if it's only set as the default one but does not include any parameters.
The test was enhanced to test this behavior better.

The parameter lookup is now better as it checks for ambiguous parameters
in all parameter sets (as PS does) and tries real names and aliases first,
which is quicker.

Last but not least, RuntimeExceptions now always create an ErrorRecord if
requested for better error handling.

All existing reference tests work now also with Pash.