Adjustments of tests; Bugfixing
First of all, a TestHostRawUserInterface was added as its WindowSize property
is needed by the new output system.

More important:
Because of the new output system, it is important now to distinguish between
objects and their string representation. As the old output system always
just called ToString() several tests had to be adjusted to use either
a specific property for comparison or explicitly their string representation
(by using ToString() explicitly).
Moreover, Path related cmdlets had to be adjusted as they returned Path
objects, not strings as in PS, which lead to inconvinient results.