Fixed bugs
Several bug fixes:
BufferSize has to be used for determining the output width instead of
the WindowWidth

The List format didn't include a newline between multiple objects.

Also the ListFormats used the same properties for each object, which
isn't necessary. Therefore the TableFormatProcessor itself cares about
which properties will printed and which  won't.

Added tests that check the detected bugs.