Distinguish whether using interative IO or not
Some features, as reading from the user (Readline in LocalHostUserInterface)
might only be available when Pash is used within a console, like the PashConsole/FullHost.
However, it might also just be used as a framework, which should not depend on user
interaction but throw an exception instead, which can be catched and reacted on.