Better support for Hashtables/Enumeratables
LanguagePrimitives provide a unique way to access an enumerable/enumerator.
Therefore all the (nearly identical) implementations in different classes could be
removed.

One important change is that only IList are seen as enumerables. For most objects as
arrays and lists, Pashs behavior won't change. But strings, Dictionaries, Hashatbles,
etc. have to be enumerated explicitly (via .GetEnumerator()). This matches Powershell's
behavior and allows Hashtables to be used in cmdlets dircetly, without being enumerated
to DictionaryEntries.

Last but not least, the FormatManager had to be adjusted to enumerate *all* objects, not
only ILists, so also Dictionaries are formatted correctly.
Unrolly one level of an input object while formatting also matches PS behavior.