Changed the enumeration behavior; Simplified tests
Obviously it was a bad idea to only enumerate ILists. That's why I changed it
back s.t. everything is enumerated but strings and IDictionaries which seems
to fit PS' behavior better.
The previously added tests (which didn't work before) make sure enumerables
are enumerated while dictionaries and Hashtables aren't.

Also the Formatting tests were simplified. The tests depended on a concrete
order of the object's properties. However, the real order is abritrary, so
the tests are more robust