Report compilation errors when using Add-Type
All compilation errors are written to error stream and are available
via the $error variable.
The Add-Type cmdlet is terminated if there are compilation errors.

There are some problems here. Pash currently only shows the error in
the console from the exception that terminates the cmdlet.
PowerShell shows the exception error and all the compilation errors.
The error from the exception results in a duplicate error record. If
the cmdlet is changed so it does not terminate with an exception then
the compiler errors are displayed in the console.