Finished joining most unary operator expressions
With the new implementation Pash starts significantly faster than before
(only ~39% of the time for starting and exiting).

All expressions with unary operators have been joined in one expression to achieve this.
The cast expression, however, had to stay there as a separate rule. Although I thought
that joining it wouldn't make a semantical difference for the grammar, a double-cast
like [int][string] wouldn't be parsed anymore.